refactor: move weather data functions into separate modules
