+++ include/wunderground.h
@@ -1,0 +1,6 @@
+#ifndef WUNDERGROUND_H
+#define WUNDERGROUND_H
+
+void sendToWeatherUnderground(float temperature_f, int humidity, float baromin, float dewpoint_f);
+
+#endif // WUNDERGROUND_H
